Process date file frontmatter first to improve date stamping.
If a file doesn't have a date in the frontmatter, the code
looks through the backlinks to find a date there. In many
cases, the files without dates are the ones created automatically
through backlinks from a date-based note (e.g. 2020-04-26.md).
Those notes provide a good hint about what the date should be,
so we process frontmatter from those files first.

Fixes #2.